# openapi-comparison/0_OpenAPI/python-flask-server-generated/swagger_server/models/book_details.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class BookDetails(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, book_id: str=None, title: str=None, author: str=None, publish_year: int=None):  # noqa: E501
        """BookDetails - a model defined in Swagger

        :param book_id: The book_id of this BookDetails.  # noqa: E501
        :type book_id: str
        :param title: The title of this BookDetails.  # noqa: E501
        :type title: str
        :param author: The author of this BookDetails.  # noqa: E501
        :type author: str
        :param publish_year: The publish_year of this BookDetails.  # noqa: E501
        :type publish_year: int
        """
        self.swagger_types = {
            'book_id': str,
            'title': str,
            'author': str,
            'publish_year': int
        }

        self.attribute_map = {
            'book_id': 'bookId',
            'title': 'title',
            'author': 'author',
            'publish_year': 'publishYear'
        }
        self._book_id = book_id
        self._title = title
        self._author = author
        self._publish_year = publish_year

    @classmethod
    def from_dict(cls, dikt) -> 'BookDetails':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The BookDetails of this BookDetails.  # noqa: E501
        :rtype: BookDetails
        """
        return util.deserialize_model(dikt, cls)

    @property
    def book_id(self) -> str:
        """Gets the book_id of this BookDetails.

        An identifier of book  # noqa: E501

        :return: The book_id of this BookDetails.
        :rtype: str
        """
        return self._book_id

    @book_id.setter
    def book_id(self, book_id: str):
        """Sets the book_id of this BookDetails.

        An identifier of book  # noqa: E501

        :param book_id: The book_id of this BookDetails.
        :type book_id: str
        """

        self._book_id = book_id

    @property
    def title(self) -> str:
        """Gets the title of this BookDetails.

        The title of a book  # noqa: E501

        :return: The title of this BookDetails.
        :rtype: str
        """
        return self._title

    @title.setter
    def title(self, title: str):
        """Sets the title of this BookDetails.

        The title of a book  # noqa: E501

        :param title: The title of this BookDetails.
        :type title: str
        """

        self._title = title

    @property
    def author(self) -> str:
        """Gets the author of this BookDetails.

        The author name  # noqa: E501

        :return: The author of this BookDetails.
        :rtype: str
        """
        return self._author

    @author.setter
    def author(self, author: str):
        """Sets the author of this BookDetails.

        The author name  # noqa: E501

        :param author: The author of this BookDetails.
        :type author: str
        """

        self._author = author

    @property
    def publish_year(self) -> int:
        """Gets the publish_year of this BookDetails.

        The year the book was published  # noqa: E501

        :return: The publish_year of this BookDetails.
        :rtype: int
        """
        return self._publish_year

    @publish_year.setter
    def publish_year(self, publish_year: int):
        """Sets the publish_year of this BookDetails.

        The year the book was published  # noqa: E501

        :param publish_year: The publish_year of this BookDetails.
        :type publish_year: int
        """

        self._publish_year = publish_year