# openapi-comparison/0_OpenAPI/python-flask-server-generated/swagger_server/models/book_list_response.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.book_details import BookDetails  # noqa: F401,E501
from swagger_server import util


class BookListResponse(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, books: List[BookDetails]=None):  # noqa: E501
        """BookListResponse - a model defined in Swagger

        :param books: The books of this BookListResponse.  # noqa: E501
        :type books: List[BookDetails]
        """
        self.swagger_types = {
            'books': List[BookDetails]
        }

        self.attribute_map = {
            'books': 'books'
        }
        self._books = books

    @classmethod
    def from_dict(cls, dikt) -> 'BookListResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The BookListResponse of this BookListResponse.  # noqa: E501
        :rtype: BookListResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def books(self) -> List[BookDetails]:
        """Gets the books of this BookListResponse.

        A list of book details  # noqa: E501

        :return: The books of this BookListResponse.
        :rtype: List[BookDetails]
        """
        return self._books

    @books.setter
    def books(self, books: List[BookDetails]):
        """Sets the books of this BookListResponse.

        A list of book details  # noqa: E501

        :param books: The books of this BookListResponse.
        :type books: List[BookDetails]
        """

        self._books = books